Extract shared Responses handlers, xAI adapters
Refactor provider code to centralize the OpenAI Responses wire format and add xAI-specific adapters.

- Add a neutral Responses Text handler (Providers\Responses\Handlers\Text) that implements the common Responses API flow (text, streaming, structured output, token counting).
- Move MediaResolver, ResponseParser, and ToolMapper into the Providers\Responses namespace and update imports/usages.
- Make OpenAi Text handler extend the neutral Responses text handler and keep only the OpenAI-Organization header behavior.
- Update ResponsesDriver to use the neutral Responses classes.
- Add xAI-specific ToolMapper and ResponseParser subclasses to handle xAI divergences (numeric call_id fallback and reasoning content location).
- Adjust/rename tests to match the new Responses namespace and add tests for xAI parser/mapper behavior.
- Bump CHANGELOG to v3.6.0.

This centralizes Responses-format logic so multiple providers (OpenAI, xAI, and Responses-compatible proxies) can share behavior while allowing small provider-specific overrides.
