+++ tests/Unit/Providers/OpenAi/BatchHandlerTest.php
@@ -10,11 +10,11 @@ use Atlasphp\Atlas\Http\HttpClient;
 use Atlasphp\Atlas\Providers\OpenAi\Handlers\Batch;
 use Atlasphp\Atlas\Providers\OpenAi\Handlers\Embed;
 use Atlasphp\Atlas\Providers\OpenAi\Handlers\Text;
-use Atlasphp\Atlas\Providers\OpenAi\MediaResolver;
 use Atlasphp\Atlas\Providers\OpenAi\MessageFactory;
-use Atlasphp\Atlas\Providers\OpenAi\ResponseParser;
-use Atlasphp\Atlas\Providers\OpenAi\ToolMapper;
 use Atlasphp\Atlas\Providers\ProviderConfig;
+use Atlasphp\Atlas\Providers\Responses\MediaResolver;
+use Atlasphp\Atlas\Providers\Responses\ResponseParser;
+use Atlasphp\Atlas\Providers\Responses\ToolMapper;
 use Atlasphp\Atlas\Requests\Batch as BatchRequest;
 use Atlasphp\Atlas\Requests\BatchLine;
 use Atlasphp\Atlas\Requests\EmbedRequest;
