+++ tests/Unit/Providers/Xai/Handlers/TextHandlerTest.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 use Atlasphp\Atlas\Enums\ReasoningEffort;
 use Atlasphp\Atlas\Exceptions\ProviderException;
 use Atlasphp\Atlas\Http\HttpClient;
-use Atlasphp\Atlas\Providers\OpenAi\MediaResolver;
-use Atlasphp\Atlas\Providers\OpenAi\ResponseParser;
-use Atlasphp\Atlas\Providers\OpenAi\ToolMapper;
 use Atlasphp\Atlas\Providers\ProviderConfig;
+use Atlasphp\Atlas\Providers\Responses\MediaResolver;
 use Atlasphp\Atlas\Providers\Tools\WebSearch;
 use Atlasphp\Atlas\Providers\Xai\Handlers\Text;
 use Atlasphp\Atlas\Providers\Xai\MessageFactory;
+use Atlasphp\Atlas\Providers\Xai\ResponseParser;
+use Atlasphp\Atlas\Providers\Xai\ToolMapper;
 use Atlasphp\Atlas\RequestConfig;
 use Atlasphp\Atlas\Requests\Reasoning;
 use Atlasphp\Atlas\Requests\TextRequest;
